yaml: Factor out example_structs for unit tests
Tidy up the struct a little bit to resolve differences between the two copies:
Serialize as first thing in struct, use kNominalDouble not NAN, always have a
non-zero, non-empty default value, etc.
